make zerovec use zerofrom directly
