+++ utils/zerovec/src/yoke_impls.rs
@@ -263,73 +263,18 @@ where
     }
 }
 
-impl<'zcf, T> ZeroCopyFrom<'zcf, ZeroVec<'_, T>> for ZeroVec<'zcf, T>
-where
-    T: 'static + AsULE + ?Sized,
-{
-    #[inline]
-    fn zero_copy_from(cart: &'zcf ZeroVec<'_, T>) -> Self {
-        ZeroVec::Borrowed(cart.as_ule_slice())
-    }
-}
-
-impl<'zcf, T> ZeroCopyFrom<'zcf, VarZeroVec<'_, T>> for VarZeroVec<'zcf, T>
-where
-    T: 'static + VarULE + ?Sized,
-{
-    #[inline]
-    fn zero_copy_from(cart: &'zcf VarZeroVec<'_, T>) -> Self {
-        cart.as_slice().into()
-    }
-}
-
-impl<'zcf, 's, K, V> ZeroCopyFrom<'zcf, ZeroMap<'s, K, V>> for ZeroMap<'zcf, K, V>
-where
-    K: 'static + for<'b> ZeroMapKV<'b> + ?Sized,
-    V: 'static + for<'b> ZeroMapKV<'b> + ?Sized,
-    <K as ZeroMapKV<'zcf>>::Container: ZeroCopyFrom<'zcf, <K as ZeroMapKV<'s>>::Container>,
-    <V as ZeroMapKV<'zcf>>::Container: ZeroCopyFrom<'zcf, <V as ZeroMapKV<'s>>::Container>,
-{
-    fn zero_copy_from(cart: &'zcf ZeroMap<'s, K, V>) -> Self {
-        ZeroMap {
-            keys: K::Container::zero_copy_from(&cart.keys),
-            values: V::Container::zero_copy_from(&cart.values),
-        }
-    }
-}
-
-impl<'zcf, 's, K0, K1, V> ZeroCopyFrom<'zcf, ZeroMap2d<'s, K0, K1, V>>
-    for ZeroMap2d<'zcf, K0, K1, V>
-where
-    K0: 'static + for<'b> ZeroMapKV<'b> + ?Sized,
-    K1: 'static + for<'b> ZeroMapKV<'b> + ?Sized,
-    V: 'static + for<'b> ZeroMapKV<'b> + ?Sized,
-    <K0 as ZeroMapKV<'zcf>>::Container: ZeroCopyFrom<'zcf, <K0 as ZeroMapKV<'s>>::Container>,
-    <K1 as ZeroMapKV<'zcf>>::Container: ZeroCopyFrom<'zcf, <K1 as ZeroMapKV<'s>>::Container>,
-    <V as ZeroMapKV<'zcf>>::Container: ZeroCopyFrom<'zcf, <V as ZeroMapKV<'s>>::Container>,
-{
-    fn zero_copy_from(cart: &'zcf ZeroMap2d<'s, K0, K1, V>) -> Self {
-        ZeroMap2d {
-            keys0: K0::Container::zero_copy_from(&cart.keys0),
-            joiner: ZeroVec::zero_copy_from(&cart.joiner),
-            keys1: K1::Container::zero_copy_from(&cart.keys1),
-            values: V::Container::zero_copy_from(&cart.values),
-        }
-    }
-}
-
 #[cfg(test)]
 #[allow(non_camel_case_types)]
 mod test {
     use super::*;
     use crate::{VarZeroSlice, ZeroSlice};
 
-    #[derive(yoke::Yokeable, yoke::ZeroCopyFrom)]
+    #[derive(yoke::Yokeable, zerofrom::ZeroFrom)]
     struct DeriveTest_ZeroVec<'data> {
         _data: ZeroVec<'data, u16>,
     }
 
-    #[derive(yoke::Yokeable, yoke::ZeroCopyFrom)]
+    #[derive(yoke::Yokeable, zerofrom::ZeroFrom)]
     struct DeriveTest_VarZeroVec<'data> {
         _data: VarZeroVec<'data, str>,
     }
@@ -339,7 +284,7 @@ mod test {
         _data: &'data VarZeroSlice<str>,
     }
 
-    #[derive(yoke::Yokeable, yoke::ZeroCopyFrom)]
+    #[derive(yoke::Yokeable, zerofrom::ZeroFrom)]
     #[yoke(prove_covariance_manually)]
     struct DeriveTest_ZeroMap<'data> {
         _data: ZeroMap<'data, [u8], str>,
@@ -351,13 +296,13 @@ mod test {
         _data: ZeroMapBorrowed<'data, [u8], str>,
     }
 
-    #[derive(yoke::Yokeable, yoke::ZeroCopyFrom)]
+    #[derive(yoke::Yokeable, zerofrom::ZeroFrom)]
     #[yoke(prove_covariance_manually)]
     struct DeriveTest_ZeroMapWithULE<'data> {
         _data: ZeroMap<'data, ZeroSlice<u32>, str>,
     }
 
-    #[derive(yoke::Yokeable, yoke::ZeroCopyFrom)]
+    #[derive(yoke::Yokeable, zerofrom::ZeroFrom)]
     #[yoke(prove_covariance_manually)]
     struct DeriveTest_ZeroMap2d<'data> {
         _data: ZeroMap2d<'data, u16, u16, str>,
